application/requirements.txt
unidecode==1.3.8
pydantic==2.8.2
fastapi[all]==0.111.1
alembic==1.11.3
SQLAlchemy[asyncio]==2.0.28
aiohttp==3.9.5
asyncpg==0.28.0
XlsxWriter==3.1.0
argon2-cffi==21.3.0
PyJWT==2.7.0
numpy
scipy==1.11.4
pandas==2.0.1
Faker==19.3.1
python-dateutil==2.8.2
websockets==12.0
openpyxl==3.1.2
xmltodict==0.13.0
xsdata[cli,lxml,soap]==24.2.1
microsoft-kiota-abstractions==1.9.0
msgraph-sdk==1.2.0
pyxlsb==1.0.10
aiofiles==24.1.0
xlrd==2.0.1
requests-toolbelt==1.0.0
torch
torchvision
easyocr==1.7.2
playwright==1.51.0